--- backend/src/main/java/io/metersphere/base/domain/UiModule.java
@@ -1,29 +1,0 @@
-package io.metersphere.base.domain;
-
-import java.io.Serializable;
-import lombok.Data;
-
-@Data
-public class UiModule implements Serializable {
-    private String id;
-
-    private String projectId;
-
-    private String name;
-
-    private String protocol;
-
-    private String parentId;
-
-    private Integer level;
-
-    private Long createTime;
-
-    private Long updateTime;
-
-    private Double pos;
-
-    private String createUser;
-
-    private static final long serialVersionUID = 1L;
-}
