+++ backend/src/main/resources/db/migration/V115__1.20__release.sql
@@ -8,23 +8,6 @@ VALUES
 		1
 	);
 
-CREATE TABLE `ui_module`
-(
-    `id`          varchar(50) NOT NULL COMMENT 'ID',
-    `project_id`  varchar(50) NOT NULL COMMENT 'Project ID this node belongs to',
-    `name`        varchar(64) NOT NULL COMMENT 'Node name',
-    `protocol`    varchar(64) NOT NULL COMMENT 'Node protocol',
-    `parent_id`   varchar(50)  DEFAULT NULL COMMENT 'Parent node ID',
-    `level`       int(10) DEFAULT '1' COMMENT 'Node level',
-    `create_time` bigint(13) NOT NULL COMMENT 'Create timestamp',
-    `update_time` bigint(13) NOT NULL COMMENT 'Update timestamp',
-    `pos`         double       DEFAULT NULL,
-    `create_user` varchar(100) DEFAULT NULL,
-    PRIMARY KEY (`id`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
-
-
-
 CREATE TABLE `ui_scenario_module`
 (
     `id`           varchar(50) NOT NULL COMMENT 'ui scenario node ID',
@@ -35,7 +18,7 @@ CREATE TABLE `ui_scenario_module`
     `create_time`  bigint(13) NOT NULL COMMENT 'Create timestamp',
     `update_time`  bigint(13) NOT NULL COMMENT 'Update timestamp',
     `pos`          double       DEFAULT NULL,
-    `create_useru` varchar(100) DEFAULT NULL,
+    `create_user` varchar(100) DEFAULT NULL,
     PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
